+++ pythonFiles/include/blender_vscode/communication.py
@@ -1,21 +1,26 @@
+import random
+import threading
 import time
+from functools import partial
 from typing import Callable, Dict
 
-import flask
 import debugpy
-import random
+import flask
 import requests
-import threading
-from functools import partial
-from .utils import run_in_main_thread
+
 from .environment import blender_path, scripts_folder, python_path
+from .utils import run_in_main_thread
+from . import log
+
+LOG = log.getLogger()
 
 EDITOR_ADDRESS = None
 OWN_SERVER_PORT = None
 DEBUGPY_PORT = None
 
-server = flask.Flask("Blender Server")
-post_handlers = {}
+SERVER = flask.Flask("Blender Server")
+log.configure_flask_log(SERVER)
+POST_HANDLERS = {}
 
 
 def setup(address: str, path_mappings):
@@ -27,9 +32,9 @@ def setup(address: str, path_mappings):
 
     send_connection_information(path_mappings)
 
-    print("Waiting for debug client.")
+    LOG.info("Waiting for debug client.")
     debugpy.wait_for_client()
-    print("Debug client attached.")
+    LOG.info("Debug client attached.")
 
 
 def start_own_server():
@@ -39,7 +44,7 @@ def start_own_server():
         while True:
             try:
                 port[0] = get_random_port()
-                server.run(debug=True, port=port[0], use_reloader=False)
+                SERVER.run(debug=True, port=port[0], use_reloader=False)
             except OSError:
                 pass
 
@@ -71,30 +76,38 @@ def start_debug_server():
 #########################################
 
 
-@server.route("/", methods=["POST"])
+@SERVER.route("/", methods=["POST"])
 def handle_post():
     data = flask.request.get_json()
-    print("Got POST:", data)
+    LOG.debug(f"Got POST: {data}")
 
-    if data["type"] in post_handlers:
-        return post_handlers[data["type"]](data)
+    if data["type"] in POST_HANDLERS:
+        return POST_HANDLERS[data["type"]](data)
+    else:
+        LOG.warning(f"Unhandled POST: {data}")
 
     return "OK"
 
 
-@server.route("/", methods=["GET"])
+@SERVER.route("/", methods=["GET"])
 def handle_get():
     data = flask.request.get_json()
-    print("Got GET:", data)
+    LOG.debug(f"Got GET: {str(data)}")
 
     if data["type"] == "ping":
         pass
+    elif data["type"] == "complete":
+        from .blender_complete import complete
+
+        return {"items": complete(data)}
+    else:
+        LOG.warning(f"Unhandled GET: {data}")
     return "OK"
 
 
 def register_post_handler(type: str, handler: Callable):
-    assert type not in post_handlers, post_handlers
-    post_handlers[type] = handler
+    assert type not in POST_HANDLERS, POST_HANDLERS
+    POST_HANDLERS[type] = handler
 
 
 def register_post_action(type: str, handler: Callable):
@@ -123,7 +136,7 @@ def send_connection_information(path_mappings: Dict):
 
 
 def send_dict_as_json(data):
-    print("Sending:", data)
+    LOG.debug(f"Sending: {data}")
     requests.post(EDITOR_ADDRESS, json=data)